// shared-logic/model-user/model-user-member-v2.d.ts
export declare class ModelUserMemberV2 {
    id: string;
    userFullName: string;
    userMobileNumber: string;
    idAdmin: string;
    listOfPhotoProfilePublicId: string[];
    listOfCloudinaryImages: any[];
    timestamp: any;
    constructor(id?: string, userFullName?: string, userMobileNumber?: string, idAdmin?: string, listOfPhotoProfilePublicId?: string[], listOfCloudinaryImages?: any[], timestamp?: any);
}